Change Retrofit calls to use new notifications
This commit modifies the ErrorHandlingCallback, which was used for the
Retrofit calls, to use the new ErrorNotification design for delivering
errors.

The ErrorNotification interface has been changed to an abstract class,
in order to define a utility showError() method that takes the error as
a parameter, determines the appropriate error message strings and
callbacks, and calls through to the actual implementation with those
parameters.

The implementations have also been updated to describe their actual UI,
instead of just describing the query case, since it seems that the type
of notification to display can depend on factors other than just the
query type (such as content area).

All the implementations have been updated appropriately, to use the
dialog notification if they were invoked as a result of a specific user
action, the overlay notification for queries loading the data to be
solely displayed on the screen, and Snackbar notification for data
loading that was not to be displayed all over the screen.